+++ elrond-codec-derive/src/top_en_derive.rs
@@ -115,9 +115,9 @@ fn top_encode_method_bodies(
 				data_enum.variants.len() < 256,
 				"enums with more than 256 variants not supported"
 			);
-			let variant_top_encode_snippets = variant_top_encode_snippets(&name, &data_enum);
+			let variant_top_encode_snippets = variant_top_encode_snippets(name, data_enum);
 			let variant_top_encode_or_exit_snippets =
-				variant_top_encode_or_exit_snippets(&name, &data_enum);
+				variant_top_encode_or_exit_snippets(name, data_enum);
 
 			let top_encode_body = quote! {
 				match self {