+++ elrond-wasm/src/storage/mappers/map_storage_mapper.rs
@@ -93,7 +93,7 @@ where
 	/// Gets a reference to the value in the entry.
 	pub fn get(&self, k: &K) -> Option<V> {
 		if self.keys_set.contains(k) {
-			return Some(self.get_mapped_storage_value(&k));
+			return Some(self.get_mapped_storage_value(k));
 		}
 		None
 	}